requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
commonmark==0.9.1
idna==3.3
Pygments==2.11.2
python-dotenv==0.20.0
requests==2.27.1
rich==12.2.0
urllib3==1.26.9
